+++ src/xyz/oihana/schema/places/JobSite.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace xyz\oihana\schema\places;
+
+/**
+ * Represents a physical location where a specific project or construction work is executed.
+ *
+ * A JobSite is a specialized type of Site, typically temporary and project-driven,
+ * where construction, maintenance, or other field-based operations take place.
+ * It is generally defined by its operational purpose and the duration of the project.
+ *
+ * @author  Marc Alcaraz (eKameleon)
+ * @package xyz\oihana\places
+ * @since   1.0.2
+ */
+class JobSite extends Site
+{
+
+}