+++ CHANGELOG.md
@@ -13,6 +13,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - Adds the JWTAlgorithm constant class
 - Adds ItemAvailability
 - Adds the PostalAddress::extendedAddress property (new standard property in https://schema.org/PostalAddress)
+- Adds the xyz\oihana\schema\places namespace
+  - Adds the Site, JobSite, Office, Warehouse classes
 
 ## [1.0.1] - 2025-10-30